Create teamordering
Team ordering You have the results of a completed round-robin tournament in which n
teams played each other once. Each game ended either with a victory for one of the
teams or with a tie. Design an algorithm that lists the teams in a sequence so that every
team did not lose the game with the team listed immediately after it.